Change getComputerChoice to return number instead of string
Writing individual cases for each round of the game will add
unnecessary bloat and repetition to the script. If rock is assigned
a value of 0, paper a value of 1, and scissors a value of 2, then
a single conditional block with the condition (playerSelection > computerSelection)
can take care of most of the cases.