Modify last condition block in function playRound
Changing to (computerSelection > playerSelection) to be more verbose
about what type of condition will trigger a victory for the computer.
These conditional statements are only possible because getComputerChoice
was reconfigured to return a number instead of a string. By using numbers
in the conditional blocks instead of strings, many of the repetitive scenarios
that can happen with each round can easily be accounted for because scissors
will always beat paper like how two will always be larger than one, and so on.