Create function game
Function does not take ties into account, as for loop sets the terms
of how many rounds are initially played. If invoking game plays five rounds by
taking user input and calling playRound, whose purpose is solely to
compute the results of a round and return it in a string, then ties can
not be accounted for in the proposed sense (playRound should be called
again with new user and computer inputs) because it is only responsible
for returning results and nothing else (one function, one action). Thus,
in the case of a tie, game dismisses it in the for loop and continues to
the next iteration.