Change for loop in func. game to while loop
The game should be played according to an input that specifies
the number of rounds (in this case, it is hard-coded to be five
rounds). However, just because five rounds should be played, it
does not mean that in the case of a tie or invalid user input that
it should be accounted for as a round played. Therefore, using a
while loop is more appropriate because although it is unsure how
many total rounds will be played (we can skip the iterations in
the loop where we run into an error case), we can specify that we
want 'x' number of rounds to be played (in the condition).