Add conditional in func. playRound to handle invalid rounds explicitly
A new func. displayRoundResults will be made to reduce the amount of
console statements that are logged to the console every time a round
is played. In order for it to work as intended, it needs an input 'result'
that tells it what the results of the current round was (tie, player win,
computer win, etc). Having playRound return a string for invalid rounds will
help with this functionality.