Combine tie and error conditionals in func. game
The func. displayRoundResults takes 'result' as a parameter and logs
an appropriate message via checking the value through a conditional statement
in its own function block. This means that it does not matter whether or not
the result was a tie or invalid when checking inside the func. game, thus it
can be reduced to an else block for cleanliness.