Fix formatting of results text in UI textbox
Using escape characters to separate text in a string
literal and using the string literal to set the text content of
a DOM element does not work, expectedly because of the way the
textContent property formats values passed to it. Thus, the CSS
property 'white-space' was added with a value of 'pre-line' to
formally instruct the textContent property to wrap text on line breaks.
(source: Stack Overflow, https://stackoverflow.com/questions/9980416/how-can-i-insert-new-line-carriage-returns-into-an-element-textcontent)