Add feature to remove UI button functionality
After the game has finished, the buttons should no longer
output any messages or make any other changes to the UI.
However, the buttons continue to display the winner message
because of a condition that specified it to do so if the
roundCounter variable value was equal to the numOfRounds variable
value. This bug could easily be corrected by simply removing
the event listener on each button after the game finishes, which is
why all the code in the anonymous function that was running in the
addEventListener block was moved to a new function declaration
'computePlayerChoice' so that a callback function could be directly
referenced in the removeEventListener method.