Fix feature to remove UI button functionality
The line of code that was written to remove the UI button
functionality by removing the event listener from the button that
triggered it (that is, just one of the choice buttons) is faulty
because the event listener that displays the winner results is still
active in the other buttons. Thus, removing them all using the forEach
array method every time the buttons are pressed WHEN the round counter
value is equal to the number of rounds ensures that the functionality of
the buttons is removed from every button after the game has concluded.