Change condition check for value of rounds input
The previous condition only checked if the rounds input
had a value that returned true ( if (numOfRoundsInput.value) ).
This means that the conditional would always accept any value as
a valid number of rounds as long as it does not return false, for
example a string of characters. By parsing the value into an integer
first, then checking if it is a number using the isInteger method from
the Number object, values that are actual number values get accepted
while string literals with letter characters are not.