Rename function 'game' to 'checkRoundResults'
Function 'game' is too undescriptive and poorly named for
the functionality it brings to this program. It is used each time
the player presses one of the UI buttons, which fires an event that
triggers its function's code. The code checks the results of the round
and does calculations accordingly.